+++ mlcheck/utils/validators.py
@@ -2,4 +2,10 @@
 
 
 def is_dataframe(obj):
-    return False
+    """Return True if the object is a pandas DataFrame."""
+    try:
+        import pandas as pd
+    except ImportError:
+        return False
+
+    return isinstance(obj, pd.DataFrame)